tractR
TractR is a mobile application that aims to allow farmers to till their land through renting tractors only when they need to till their land. Tillage is the agricultural preparation of soil by as digging, stirring, and overturning. In order to till land that is great in size, using a tractor to till the most efficient and effective method available.